Add all files

+++ src/main/java/com/taskmanagment/digi/entities/Project.java
@@ -1,13 +1,8 @@
 package com.taskmanagment.digi.entities;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
+import lombok.*;
 import java.util.List;
-
 
 @Data
 @Entity
@@ -17,18 +12,15 @@ import java.util.List;
 public class Project {
 
     @Id
-    @GeneratedValue(generator = "projectIdGen",strategy = GenerationType.AUTO )
+    @GeneratedValue(generator = "projectIdGen", strategy = GenerationType.AUTO)
     @Column(name = "PROJECT_ID")
-        private int projectId;
+    private Long projectId;
 
     @Column(name = "PROJECT_NAME")
-        private String name;
+    private String name;
 
     @OneToMany(mappedBy = "project", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Task> tasks;
-
-
-
 
 
 }
